Remove most fake changes when saving scenes
Previously, saving the level scene would always add changes due to the
flying fish's animations and the end boat WaveMover behaviour. The
flying fishes now have a RESET track that is reverted to when saving,
and the WaveMover script has a variable to control wether it should run
in the editor, to avoid these issues. However, the tilemap still creates
fake changes when closing and reopening the scene, with signal
connections being remade. For now I see no way to fix that.